удалён лишний код
